apps/growatt-api-explorer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "growatt-api-explorer"
version = "0.1.0"
description = "Exploration tool for Growatt Cloud API - gather docs, test endpoints, validate assumptions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "growattServer>=1.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
